Refactor: Simplify entity by removing unused setter methods
Sadly I don't think we can get rid of the getters because they are called by the symphony serializer. And we can't just make the properties public because that won't work with Doctrine proxy objects.

Untested code. But this should also now make it easier to write tests, since the constructor now tells you what properties a comment needs, and you don't have to remember to call multiple setter functions.